backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
face-recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
werkzeug==3.0.3
python-dotenv==1.0.1
pytest==8.3.3
